ml/requirements.txt
## Standard science stack
h5py
numpy>=1.25.0,<=1.26.4
scipy

## Standard utils
tqdm

## Deep learning
einops
git+https://github.com/dgasmith/opt_einsum.git@1a984b7b75f3e532e7129f6aa13f7ddc3da66e10 # Un-released latest master (pinned for repro); the latest released version 3.3.0 has bugs in shape-based contraction path computations (used for einsum)
torch==2.2.0

## Experiment management
hydra-core

## Project-specific: general
clifford==1.4.0
numba==0.58.1 # otherwise has conflicts
